exhuastiveness: split functionality into respective `ops` files
